Initial navigation and routing implementation
Used the same approach that I previously worked with for the Elm Orlando
site. Seems like there should be an easier approach than this though.
Might look into some alternative approaches if they exist. Would be nice
to use a "slug" instead of an ID in the URLs for specific feels.